pyproject.toml
[project]
name = "x-newsletter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "tweepy>=4.15.0",
]
